backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Miguel Asin",email = "miguel.asin98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
httpx = "^0.28.1"
beautifulsoup4 = "^4.13.4"
pydantic-settings = "^2.9.1"
loguru = "^0.7.3"
slowapi = "^0.1.9"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
pillow = "^11.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
